Moved to and from methods from Rotation to Quaternion
